Move `HierarchyNodeKey` and `HierarchyNodeIdentifier` out of `HierarchyNode.ts` into separate source files
