+++ packages/hierarchies/src/hierarchies/HierarchyNodeIdentifier.ts
@@ -1,0 +1,51 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { InstanceKey } from "@itwin/presentation-shared";
+
+/**
+ * An identifier that can be used to identify either an ECInstance or a custom node.
+ *
+ * This is different from `HierarchyNodeKey` - the key can represent more types of nodes and,
+ * in case of `InstancesNodeKey`, contains information about all instances the node represents.
+ * `HierarchyNodeIdentifier`, on the other hand, is used for matching a node, so it only needs
+ * to contain information about a single instance or custom node key.
+ *
+ * @beta
+ */
+export type HierarchyNodeIdentifier = InstanceKey | { key: string };
+
+/** @beta */
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export namespace HierarchyNodeIdentifier {
+  /** Checks whether the given identifier is an instance node identifier */
+  export function isInstanceNodeIdentifier(id: HierarchyNodeIdentifier): id is InstanceKey {
+    return !!(id as InstanceKey).id;
+  }
+
+  /** Checks whether the given identifier is a custom node identifier */
+  export function isCustomNodeIdentifier(id: HierarchyNodeIdentifier): id is { key: string } {
+    return !!(id as { key: string }).key;
+  }
+
+  /** Checks two identifiers for equality */
+  export function equal(lhs: HierarchyNodeIdentifier, rhs: HierarchyNodeIdentifier) {
+    if (isInstanceNodeIdentifier(lhs) && isInstanceNodeIdentifier(rhs)) {
+      return lhs.className === rhs.className && lhs.id === rhs.id;
+    }
+    if (isCustomNodeIdentifier(lhs) && isCustomNodeIdentifier(rhs)) {
+      return lhs.key === rhs.key;
+    }
+    return false;
+  }
+}
+
+/**
+ * A path of hierarchy node identifiers, typically used to describe a path from root down
+ * to specific node deep in the hierarchy.
+ *
+ * @beta
+ */
+export type HierarchyNodeIdentifiersPath = HierarchyNodeIdentifier[];
